backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
easyocr==1.7.0
opencv-python==4.8.1.78
Pillow==10.1.0
PyPDF2==3.0.1
PyMuPDF==1.23.14
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
aiofiles==23.2.1
python-socketio==5.10.0
slowapi==0.1.9
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
spacy==3.7.2
websockets==12.0
numpy>=1.26.0
scikit-learn==1.3.2
transformers==4.35.2
torch==2.2.2